deployment/youtube_download/requirements.txt
tqdm
pandas
yt-dlp
boto3
python-dotenv